// FRONTEND/chatbot-admin-panel/src/services/api/schemas/databaseConnectionUpdate.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Mi Chatbot IA Personalizable
 * Un backend API en Python (FastAPI) para un chatbot avanzado con IA, configurable, capaz de conectarse a diversas fuentes de contexto y con gestión de permisos para administradores.
 * OpenAPI spec version: 0.2.0
 */
import type { DatabaseConnectionUpdateName } from './databaseConnectionUpdateName';
import type { DatabaseConnectionUpdateDescription } from './databaseConnectionUpdateDescription';
import type { DatabaseConnectionUpdateDbType } from './databaseConnectionUpdateDbType';
import type { DatabaseConnectionUpdateHost } from './databaseConnectionUpdateHost';
import type { DatabaseConnectionUpdatePort } from './databaseConnectionUpdatePort';
import type { DatabaseConnectionUpdateDatabaseName } from './databaseConnectionUpdateDatabaseName';
import type { DatabaseConnectionUpdateUsername } from './databaseConnectionUpdateUsername';
import type { DatabaseConnectionUpdatePassword } from './databaseConnectionUpdatePassword';
import type { DatabaseConnectionUpdateExtraParams } from './databaseConnectionUpdateExtraParams';

export interface DatabaseConnectionUpdate {
  name?: DatabaseConnectionUpdateName;
  description?: DatabaseConnectionUpdateDescription;
  db_type?: DatabaseConnectionUpdateDbType;
  host?: DatabaseConnectionUpdateHost;
  port?: DatabaseConnectionUpdatePort;
  database_name?: DatabaseConnectionUpdateDatabaseName;
  username?: DatabaseConnectionUpdateUsername;
  password?: DatabaseConnectionUpdatePassword;
  extra_params?: DatabaseConnectionUpdateExtraParams;
}